[RET-1960] Adding version 2 of retreaver.js
The goal is to integrate with true call and be able to experiment
with integration without risk of breaking existing publisher websites.

Moving v1 from src/retreaver to src/v1/retreaver.

There is a redundancy in the code, but in the same time we reduce the
risk by starting from a fresh copy.
Also if we want to downport something to v1 it will keep the same
process as it is now